Move hist smoothing code into mixin class
