+++ microfit/histogram/__init__.py
@@ -1,4 +1,5 @@
 from .binning import *
 from .histogram import *
+from .hist_smoothing import *
 from .histogram_generator import *
 from .run_hist_generator import *
